+++ CHANGELOG.md
@@ -24,6 +24,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - basic tests
   - bootstrap tests
   - citation tests
+  - editor:
+    - monaco: `MonacoEditorManager`, `edit_file_with_monaco`
+    - server: `app`, `ServerControl`
   - tools:
     - package_manager: `GitSolver`, `PIPInstaller`, `fetch_gist_file`, `fetch_gist_json`, `filter_sensitive_data`, `get_github_repo_tags`, `run_command`, `solve_installation_config`
     - customized_widgets: `real_bool`
@@ -33,6 +36,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - renamed:
   - `CitableModules` -> `CitableModuleAbstract`: Explicitly named as an abstract class.
+- moved:
+  - `src/REvoDesign/citations/CitationManager.py` -> `src/REvoDesign/citations/citation_manager.py`
+  - `tests/citations/test_CitationManager.py` -> `tests/citations/test_citation_manager.py`
 - refactored `SingletonAbstract`:
   - `singleton_init`: an abstract method for each subclass to implement. This method will be called when the subclass is initialized, helps to avoid repeated code duplication. This method is the **only** one developer should implement in subclasses to replace the original `__init__` method.
   - `derive`: derive a new singleton class from the current one.